Rename Grid OldGrid, preparation for replacing by redux
